Update json pull for date limit
When there's more than a year or so worth of transactions, the json pull can take a bit of time. Adding a start date helps limit the pull time. This update looks at the first 'odate' on the page each time a new page of data is pulled, and if it's earlier than the date fed into the function it will break and stop adding more data. NOTE - I'm kind of new to programming so apologies if my code isn't very elegant or doesn't make sense!