2-step work around
This commit add two cookies to the login step.  Providing these cookies
will bypass the need to 2-step authenticate.  I've run this code daily
for two weeks now and it hasn't failed yet.  I don't think this is the
best solution; but it works and I know others can benefit from it.